Check for signals during I/O operations.
During a long-running decoding or encoding operation, we need to
respond to SIGINT or other signals.

If fileobj is implemented in Python, this may not be a concern since
the interpreter should check for signals regularly.  But if fileobj is
a built-in file object (or another object implemented in C) then it
may never invoke PyErr_CheckSignals, and all signals would be ignored
until after Decoder.read or Encoder.write finishes.

There may be other places where it's necessary for us to explicitly
call PyErr_CheckSignals, but these two are the obvious ones.