encoder: pass data as PyBytes rather than PyMemoryView.
Using PyMemoryView_FromMemory allows fileobj.write() to directly
access the temporary buffer passed to encoder_write.  This is unsafe -
we can't be sure that the write method won't keep a reference after it
returns.

To be safe, make an immutable copy of the data instead; it'll just be
a tiny bit slower.  This is for encoding, so the overhead is
insignificant.