encoder: check for exceptions during init and finish.
If a Python callback raises an exception, we want to propagate that
exception to the caller; we don't want to replace it with a different
and probably-less-informative exception.

It is possible that, within FLAC__stream_encoder_init_stream or
FLAC__stream_encoder_finish, fileobj.write() or fileobj.seek() could
raise an exception, so we need to check for that.