encoder: add properties for encoder settings.
Various properties may need to be set on the stream encoder before
calling FLAC__stream_encoder_init_stream.  These generally follow the
pattern that libFLAC provides a getter and a setter function, and the
setter function will fail if the stream is already initialized.

The 'channels', 'bits_per_sample', and 'sample_rate' properties
describe the characteristics of the sample data, and most applications
will need to set these.

The 'total_samples_estimate' property is useful but not generally
required.

The 'streamable_subset' property is a meta-property that enables or
disables some advanced compression features.

The 'verify' property is used for debugging.

The remaining properties control the compression algorithm.  Setting
'compression_level' is shorthand for setting all the other properties
to predefined values.  As a result, libFLAC doesn't provide a function
to "get" the compression level, so the getter for 'compression_level'
simply returns the last value written.

Additionally, there's no function to "get" the apodization function
specification (or to check that the string we provided was
syntactically correct.)  So the getter for 'apodization' simply
returns the last value written, or None if the internal default was
selected.