decoder: handle None return value from readinto.
If we are reading from a non-blocking file, and no bytes are currently
available, then readinto returns None.

(In fact, this is the documented behavior for *raw* binary streams.
The Python manual claims that for *buffered* binary streams, a
BlockingIOError is raised instead, but in practice that is not true
for CPython.  It'd be wise to try to handle both possibilities.)

In this case, we want to abort the read operation and return to the
original caller.