encoder: fix -Wsign-compare warning.
The previous code gave warnings on 32-bit systems due to comparisons
between signed and unsigned integers.

The return type of PySequence_Length is Py_ssize_t, but if it returns
without an error then the result is necessarily non-negative.

The return type of FLAC__stream_encoder_get_channels is uint32_t (but
should never be greater than 8).  The type of the constant
FLAC__MAX_CHANNELS is unsigned int.

Thus, size_t should be able to handle all of these.