Link to the embedded copy of libFLAC.
To make the installation process clean and simple, we'll embed a copy
of the FLAC package in the source distribution, and by default,
compile and link this version of libFLAC into _plibflac.abi3.so.

To make this work requires listing the relevant source files (which is
pretty straightforward), and also defining some preprocessor macros
that would normally be defined either by configure (via config.h.in)
or by cmake (via config.cmake.h.in).

Most of these macros are unimportant or can be defined statically.
Some of them are architecture-specific, but with the exception of
CPU_IS_BIG_ENDIAN/WORDS_BIGENDIAN, they should only affect
optimization, not correctness.  All of the architecture-specific
macros, except for endianness, are ignored for the time being.

Alternatively, if the user wants to link with an existing external
copy of libFLAC, they can do so by defining FLAC_LIBS and FLAC_CFLAGS
environment variables when installing; for example, setting

  FLAC_CFLAGS=$(pkg-config --cflags flac)
  FLAC_LIBS=$(pkg-config --libs flac)

for a system package.