Support building on 3.13 free-threading CPython.
The free-threading interpreter doesn't currently support the limited
API (and isn't ABI-compatible with the GIL interpreter), so disable
the limited API in this case.

This should allow the module to be compiled - if you actually load the
module (in the current pre-release version), it will show a warning
and enable the GIL.  Further work is needed to make the objects
actually thread-safe.