Move macro definitions from setup.py into src/flac/config.h.
libFLAC expects various macros to be defined.  Some of those macros
conflict with what Python.h expects.

In particular, beginning with CPython 3.12, Python.h expects that if
the system is little-endian, WORDS_BIGENDIAN is undefined, whereas
libFLAC expects that if the system is little-endian, WORDS_BIGENDIAN
is defined with a value of zero.

Python.h relies on lots of other non-"Py"-namespaced macros, too.
It's just not a great idea to mix namespaces like this.

So instead, put all the FLAC macro definitions in a separate config.h
(based on config.cmake.h.in), which will be used when building libFLAC
(since we define HAVE_CONFIG_H) and not when building the Python API
bits.

For things that are actually system or build-dependent, have setup.py
define them in the PLIBFLAC_ namespace, and have config.h translate
appropriately.

Enabling CPU-specific optimizations is still to be done.