Add macros for managing thread state.
We want to be able to split up work among multiple threads, and allow
other threads to do work while we are busy encoding or decoding.

When one of our Python methods is called, the current thread is
holding the global interpreter lock.  We want to release that lock
while we're busy with encoding/decoding, then reacquire the lock
before returning from the method.  The BEGIN_PROCESSING and
END_PROCESSING macros implement this pattern.

Contrariwise, when one of our callbacks is called by libFLAC,
then (assuming we used BEGIN_PROCESSING) we need to reacquire the lock
before calling any Python functions.  The BEGIN_CALLBACK and
END_CALLBACK macros implement this pattern.

We're assuming that libFLAC invokes the callbacks within the same
thread from which libFLAC was originally called, which means these
macros can be exactly symmetrical.

It's important to note, however, that because we don't know exactly
which callbacks will be invoked and in what order, this must be
all-or-nothing: it's imperative to apply BEGIN/END_PROCESSING around
every libFLAC function call that might invoke callbacks, and
BEGIN/END_CALLBACK around every Python call within each of those
callbacks.