decoder: directly read from the underlying file descriptor.
If we are reading from a real file, it's much faster to access the
file descriptor directly, using read() and lseek() in C, rather than
using the Python object methods.  However, we only want to perform
this optimization if it doesn't change the results.

Although all Python file objects nominally have a 'fileno' method, we
can't rely on that method being "correct".  For example, file objects
returned by the gzip module will claim to have a file descriptor, but
reading from that file descriptor is very different than reading from
the file object.

So we only want to use fileno if the file object is actually an
instance of one of the built-in I/O classes:
- io.FileIO (returned by open(path, 'rb', buffering=0))
- io.BufferedReader wrapping an io.FileIO (open(path, 'rb'))
- io.BufferedRandom wrapping an io.FileIO (open(path, 'rb+'))
and furthermore, in the case of buffered files, only if the file is
seekable, so that we can ensure the buffer is flushed.