decoder: retry read() after EINTR.
If a read operation is interrupted by a signal, and the signal is
handled in Python (without raising an exception), then we want to
retry the read.

The code here is in essence the same as what CPython does for all
blocking system calls, including 'os.read'.  From what I can tell, the
standard approach to signal handling in CPython appears to be
fundamentally buggy and there's no way for us to avoid that - if it
matters, applications will need to work around it using
'signal.set_wakeup_fd' or similar.