decoder: release GIL while copying decoded sample arrays.
By keeping pointers to the output sample buffers (not just keeping
pointers to the Python objects), we can avoid calling Python APIs
within Decoder_write in most cases, and only acquire the GIL when we
need to allocate memory or raise an exception.

write_out_samples() is now expected to be called with GIL released.