Support Python subinterpreters.
Subinterpreters allow multiple quasi-independent instances of Python
in a single process, and can be used as an alternative to
free-threading or multi-process concurrency.  (The infrastructure for
subinterpreters was added in Python 3.12, and should be fully
supported starting with Python 3.14.)

plibflac should be fully compatible with subinterpreters since it does
not use any static variables.  However, the module needs to be flagged
as being compatible, and the Py_MOD_PER_INTERPRETER_GIL_SUPPORTED flag
will be rejected by Python versions older than 3.12.

To handle this (ensuring that the same _plibflac.abi3.so can be loaded
by Python 3.11 or earlier as well as by 3.12 or later), check the
Python version when PyInit__plibflac is called, and construct the
slots array dynamically.

(Note that this is different than Py_mod_gil: we must not add the
Py_MOD_GIL_NOT_USED flag unless the plibflac module itself is compiled
using Py_GIL_DISABLED.)