encoder: Add internal num_threads property.
In FLAC 1.5.0, the stream encoder can be set to use multiple threads
for parallel processing.

libFLAC currently has an arbitrary internal limit of 64 threads (a
limit which is not exposed directly in the API.)  If the Python caller
tries to set a very large number of threads, we should simply use the
maximum number that libFLAC supports.  It's all an opportunistic
optimization, after all.

So if FLAC__stream_encoder_set_num_threads fails because the given
value is too large, then fall back to the largest value that doesn't
fail.