decoder: Avoid unnecessary locking in main decoder loop.
When decoding a large number of frames in a row, we don't need to
check for Python errors after every call to
FLAC__stream_decoder_process_single.  (If an error does occur, the
read callback should cause libFLAC to abort as soon as possible.)

By moving all of the Python API calls out of the loop, we can avoid
re-acquiring the GIL or critical-section lock on every iteration.