decoder: add internal error_callback property.
If a non-fatal error is detected in the stream, this callable will be
invoked.  This can be used to display a warning message or to raise an
exception.  (Note that if an exception is raised, this won't stop
decoding immediately, but the next time decoder_read or
decoder_read_fd is called, that function will abort.)