decoder: add 'errors' parameter, set to 'strict' by default.
If a non-fatal error is detected in the stream, raise an exception by
default.

When creating the Decoder, an 'errors' keyword argument can be
specified, by analogy to the built-in open function.  Specifying
'errors="ignore"' means to ignore the error and try to keep going;
'errors="warn"' means to log a warning message and keep going.